requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
backcall==0.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
conda==4.6.14
conda-build==3.17.8
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.7
decorator==4.4.0
filelock==3.0.10
gast==0.2.2
glob2==0.6
grpcio==1.20.1
h5py==2.9.0
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Jinja2==2.10.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
libarchive-c==2.8
lief==0.9.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mkl-fft==1.0.12
mkl-random==1.0.2
mock==3.0.5
nose==1.3.7
numpy==1.16.3
olefile==0.46
opencv-python==3.3.0.9
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pkginfo==1.5.0.1
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.6.2
ptyprocess==0.6.0
pycocotools==2.0.0
pycosat==0.6.3
pycparser==2.19
pydensecrf==1.0rc3
Pygments==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
ruamel-yaml==0.15.46
scipy==1.2.1
setproctitle==1.1.10
six==1.12.0
soupsieve==1.8
tensorboard==1.13.1
tensorboardX==1.6
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
torch==1.1.0
torch-encoding==1.0.1
torchvision==0.2.2
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
Werkzeug==0.15.2
